+++ core/src/main/java/com/terning/core/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.terning.point.ui.theme
+package com.terning.core.ui.theme
 
 import androidx.compose.ui.graphics.Color
