+++ core/src/main/java/com/terning/core/ui/theme/Theme.kt
@@ -1,6 +1,5 @@
-package com.terning.point.ui.theme
+package com.terning.core.ui.theme
 
-import android.app.Activity
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
