+++ core/src/main/java/com/terning/core/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.terning.point.ui.theme
+package com.terning.core.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
