+++ feature/build.gradle.kts
@@ -30,6 +30,12 @@ android {
     kotlinOptions {
         jvmTarget = "1.8"
     }
+    buildFeatures {
+        compose = true
+    }
+    composeOptions {
+        kotlinCompilerExtensionVersion = "1.5.0"
+    }
 }
 
 dependencies {
@@ -63,7 +69,8 @@ dependencies {
 
     // <4>
     implementation(libs.material)
-    implementation(libs.androidx.material3)
+    implementation(libs.androidx.material3.android)
+    implementation(libs.androidx.ui.tooling.preview)
 
     // <5> TestDependencies
     testImplementation(libs.junit)
